Began Version 0.1:
Created the main program file bacch.py, laid out the main and auxiliary functions.  Configuerd the argument parser and subparsers for the three major jobs bacch is to undertake in this version: building presentation documents, updating the structural file, compiling data on the current state and progress of the project.

Updated README.md with more detailed information about what bacch is meant to do.  Added a section on Current Development, noting the specific goals for Version 0.1 and in particular that the project is currently unstable.

Added a .gitignore file, blacklisting temporary files and autosaves from Emacs.