Updates to various functions covered under the struct command, previously known as the updater.
bacch struct now checks for a source directory, then reads the files in said directory.
Firstly, it determines whether the source project is written primarily in XML or reStructuredText,
then, in the event that it is reStructuredText, it checks the most recent changes against the
.struct.json file in the project root.  If the source files are more recent, it initiates a process
to update the .struct.json file - but, does not yet do anything past writing in a blank string.